add direction.py
